+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## main
+- [added] PWA
+- [added] `tienda` Export albaranes list.
+
 ## 0.1.1 - 2023-06-16
 - [fixed] Regenerate users initial migration (changed user manager).